+++ spec/integration/resource/api600/c7000/logical_enclosure/create_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::C7000::LogicalEnclosure
+klass = OneviewSDK::API600::C7000::LogicalEnclosure
 RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
-  let(:current_client) { $client_500 }
-  include_examples 'LogicalEnclosureCreateExample', 'integration api500 context', 'C7000'
+  let(:current_client) { $client_600 }
+  include_examples 'LogicalEnclosureCreateExample', 'integration api600 context', 'C7000'
 end