+++ spec/integration/resource/api600/c7000/logical_enclosure/update_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::C7000::LogicalEnclosure
+klass = OneviewSDK::API600::C7000::LogicalEnclosure
 RSpec.describe klass, integration: true, type: UPDATE do
-  let(:current_client) { $client_500 }
+  let(:current_client) { $client_600 }
   let(:encl_name) { ENCL_NAME }
-  include_examples 'LogicalEnclosureUpdateExample', 'integration api500 context', 500, 'C7000'
+  include_examples 'LogicalEnclosureUpdateExample', 'integration api600 context', 600, 'C7000'
 end