+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This release adds support to OneView Rest API version 600 for the hardware varia
    - Interconnect
    - Interconnect Type
    - Internal Link Set
+   - Logical Enclosure
    - Logical Interconnect
    - SAS Interconnect
    - SAS Interconnect Type